// homecare-backend/src/interfaces/http/dtos/patientContactDto.ts
export interface PatientContactResponseDto {
  id:                 string
  patientId:          string
  name:               string
  relationship:       string
  phone:              string
  phoneAlternative:   string | null
  email:              string | null
  livesAtCareAddress: boolean
  availabilityHours:  string | null
  isPrimary:          boolean
  createdAt:          string
}